+++ src/main/java/pl/plantoplate/REST/service/PantryService.java
@@ -83,7 +83,7 @@ public class PantryService {
      * @param email - email of user to identify his group
      * @return
      */
-    public List<ShopProduct> addProductToPantry(long productId, int amount, String email) {
+    public List<ShopProduct> addProductToPantry(long productId, float amount, String email) {
 
         Group group = userService.findGroupOfUser(email);
 
@@ -154,7 +154,7 @@ public class PantryService {
      * @param amount - new amount of product
      * @return
      */
-    public List<ShopProduct> modifyAmount(long pantryProductId, String email, int amount) {
+    public List<ShopProduct> modifyAmount(long pantryProductId, String email, float amount) {
 
         Group group = userService.findGroupOfUser(email);
